keep predict_clr login in clr.py; house keeping
